% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_and_bias.R
\name{accuracy_and_bias}
\alias{accuracy_and_bias}
\title{Accuracy and Bias Indicators}
\usage{
accuracy_and_bias(df)
}
\arguments{
\item{{df}}{{data frame containing, at least,  a column \code{Error} or \code{Residual}, and a
column \code{Demand}}}
}
\value{
A list containing elements for the following indicators:
\itemize{
\item\code{Av. Error} {average error; a bias indicator}
\item\code{Scaled Av. Err.}  {scaled average error; bias expressed in percentage}
\item\code{MAD} {mean absolute deviation / error}
\item\code{Scaled MAD} {scaled mean absolute error}
\item\code{MAPE} {mean absolute percentage error}
\item\code{MSE} {mean square error}
\item\code{Scaled MSE} {scaled mean square error}
\item\code{RMSE} {root mean square error}
\item{Scaled RMSE} {scaled root mean square error}
}
It returns \code{NULL} if input does not match requirements
}
\description{
Measures the Forecast Error and provides Accuracy and Bias Indicators
}
\examples{
 ## Forecast
 ##
 r0 <- des(dTest2, future=0)
 ##
 ## Compute accuracy and bias indicators
 ##
 accuracy_and_bias(r0)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
