% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxsProb1.R
\docType{data}
\name{AxsProb1}
\alias{AxsProb1}
\title{Example of demand data}
\format{
Numerical vector
}
\usage{
AxsProb1
}
\description{
Data based on Problem 2.1 in Axsätter's book.
}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\seealso{
AxsProb13, AxsProb9
}
\keyword{datasets}
