% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxsProb13.R
\docType{data}
\name{AxsProb13}
\alias{AxsProb13}
\title{Example of demand data}
\format{
Numerical vector
}
\usage{
AxsProb13
}
\description{
Data based on Problem 2.13 in Axsätter's book.
}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\seealso{
AxsProb1, AxsProb9
}
\keyword{datasets}
