% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxsProb9.R
\docType{data}
\name{AxsProb9}
\alias{AxsProb9}
\title{Example of demand data}
\format{
Numerical vector
}
\usage{
AxsProb9
}
\description{
Data based on Problem 2.9 in Axsätter's book.
}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
}
}
\seealso{
AxsProb1, AxsProb13
}
\keyword{datasets}
