% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brown5460am.R
\docType{data}
\name{Brown5460am}
\alias{Brown5460am}
\title{Example of US domestic airline passenger miles (in billions)}
\format{
Numerical vector
}
\usage{
Brown5460am
}
\description{
Data is based on Table 4.1 in Brown below
}
\references{
Brown, R. G. (1962, 1963) Smoothing, Forecasting, and Prediction, Prentice-Hall, London.
}
\keyword{datasets}
