% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrownTC10monyr.R, R/BrownTC16monyr.R
\docType{data}
\name{BrownTC10monyr}
\alias{BrownTC10monyr}
\title{Example of monthly international airlines passengers for the year 1949--1960}
\format{
Data frame with 13 columns, named:
\itemize{
\item\code{Month} {containing the 12 months \code{Jan} to \code{Dec} in sequence}
\item{\code{Y1949} -- \code{Y1960}} {containing montly number of passengers (in thousand)}
}

Data frame with 11 columns, named:
\itemize{
\item\code{Month} {containing the 12 months \code{Jan} to \code{Dec} in sequence}
\item{\code{Y1948} -- \code{Y1957}} {containing montly business conditions)}
}
}
\usage{
BrownTC10monyr

BrownTC10monyr
}
\description{
Data is based on Table C.10  in Brown below. Thousands are omitted.

Data is based on Table C.16  in Brown below.
}
\references{
Brown, R. G. (1962, 1963) Smoothing, Forecasting, and Prediction, Prentice-Hall, London.

Brown, R. G. (1962, 1963) Smoothing, Forecasting, and Prediction, Prentice-Hall, London.
}
\keyword{datasets}
