% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrownTC12vec.R
\docType{data}
\name{BrownTC12vec}
\alias{BrownTC12vec}
\title{Example of data expressing weekly sales of Warmdot Thermostats}
\format{
A numeric vector of length 52
}
\usage{
BrownTC12vec
}
\description{
Data is based on Table C.12  in Brown below.
}
\references{
Brown, R. G. (1962, 1963) Smoothing, Forecasting, and Prediction, Prentice-Hall, London.
}
\keyword{datasets}
