% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrownTC14monyr.R
\docType{data}
\name{BrownTC14monyr}
\alias{BrownTC14monyr}
\title{Example of monthly sales of Warmdot Filters for the year 1953--1962}
\format{
Data frame with 11 columns, named:
\itemize{
\item\code{Month} {containing the 12 months \code{Jan} to \code{Dec} in sequence}
\item{\code{Y1953} -- \code{Y1962}} {containing montly sales}
}
}
\usage{
BrownTC14monyr
}
\description{
Data is based on Table C.14  in Brown below.
}
\references{
Brown, R. G. (1962, 1963) Smoothing, Forecasting, and Prediction, Prentice-Hall, London.
}
\keyword{datasets}
