% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrownTC1vec.R
\docType{data}
\name{BrownTC1vec}
\alias{BrownTC1vec}
\title{A sample sequence of data}
\format{
A numeric vector of length 100
}
\usage{
BrownTC1vec
}
\description{
Data is based on Table C.1  in Brown below.
}
\references{
Brown, R. G. (1962, 1963) Smoothing, Forecasting, and Prediction, Prentice-Hall, London.
}
\keyword{datasets}
