% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrownWS1.R
\docType{data}
\name{BrownWS1}
\alias{BrownWS1}
\title{Example of  data of installationd}
\format{
Numerical vector
}
\usage{
BrownWS1
}
\description{
Data is based on Work Sheet No. 1 in Brown below. Data multiplied by 1000.
}
\references{
Brown, R. G. (1962, 1963) Smoothing, Forecasting, and Prediction, Prentice-Hall, London.
}
\keyword{datasets}
