% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrownWS4.R
\docType{data}
\name{BrownWS4}
\alias{BrownWS4}
\title{Example of  data of Warmdot Tgermostat Weekly Sales}
\format{
Numerical vector
}
\usage{
BrownWS4
}
\description{
Data is based on Work Sheet No. 4 in Brown below.
}
\references{
Brown, R. G. (1962, 1963) Smoothing, Forecasting, and Prediction, Prentice-Hall, London.
}
\keyword{datasets}
