% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrownWS7.R
\docType{data}
\name{BrownWS7}
\alias{BrownWS7}
\title{Example of  data of Daily Closing Prices of IBM Stock between August 8 and 19, 1960}
\format{
Numerical vector
}
\usage{
BrownWS7
}
\description{
Data is based on Work Sheet No. 7 in Brown below.
}
\references{
Brown, R. G. (1962, 1963) Smoothing, Forecasting, and Prediction, Prentice-Hall, London.
}
\keyword{datasets}
