% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition_lrtp_add.R
\name{decomposition_lrtp_add}
\alias{decomposition_lrtp_add}
\title{Time Series ecomposition based on Regression Trend Projection -- Additive Model}
\usage{
decomposition_lrtp_add(d, fp = NaN)
}
\arguments{
\item{{d}}{{a vector containing observed demand}}

\item{{fp}}{{seasonal period frequency, if any}}
}
\value{
A list containing:
\itemize{
\item\code{Seasonal}    {Seasonal additive components}
\item\code{TrendLine}   {Regression coefficients for the trend equation}
\item\code{Table} {data frame with columns:
\itemize{
\item\code{Demand} {the given demand}
\item\code{DeSeasoned} {de-seasoned demand}
\item\code{Residual} {de-seasoned and de-trended remains of demand}
\item\code{Seasonality} {the seasonal components spread over the entire horizon}
\item\code{Trend} {estimated trend of the demand}
}}
}
}
\description{
Linear regression method is applied to estimate the time series trend
and subsequently, if seasonality is present, additive seasonal
components of the de-trended series
are computed and the de-trended series is de-seasonized.
}
\details{
Requires AMProgosis::seasonal_add(), AMPrognosis::lrtp()
}
\examples{
## Load test data, for instance
##
data("dTest6")
##
## Apply additive decomposition based on linear regression
##
res<-decomposition_lrtp_add(dTest6, fp=4)
##
## Plot the time series trend
##
plot_demand_vs_forecast(data.frame(Demand=res$Table$Demand,Trend=res$Table$Trend))
##
## Plot the decomposition components
##
plot_decomposition(res$Table)

}
\seealso{
decomposition_lrtp_mult, decomposition_lrtp, decomposition_sma_add,
decomposition_sma_mult, decomposition_sma, plot_decomposition
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
