% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition_sma.R
\name{decomposition_sma}
\alias{decomposition_sma}
\title{Common Interface for Time Series Decomposition based on Symmteric Moving Average estimation of Trend}
\usage{
decomposition_sma(d, fp = NaN, model = c("additive", "multiplicative"))
}
\arguments{
\item{{d}}{{a vector containing observed demand}}

\item{{fp}}{{seasonal period frequency, if any}}

\item{{model}}{{choice of \code{"additive"} or \code{"multiplicative"}}}
}
\value{
A list containing:
\itemize{
\item\code{Seasonal}    {Seasonal additive components}
\item\code{TrendLine}   {Regression coefficients for the trend equation}
\item\code{Table} {data frame with columns:
\itemize{
\item\code{Demand} {the given demand}
\item\code{DeSeasoned} {de-seasoned demand}
\item\code{Residual} {de-seasoned and de-trended remains of demand}
\item\code{Seasonality} {the seasonal components spread over the entire horizon}
\item\code{Trend} {estimated trend of the demand}
}}
}
}
\description{
Symmetric Moving Average method is applied to estimate the time series trend
and subsequently, if seasonality is present, additive or multiplicative seasonal
components of the de-trended series
are computed and the de-trended series is de-seasonized.
}
\details{
Requires AMProgosis::decomposition_sma_add(), and AMPrognosis::decomposition_sma_mult()
}
\examples{
## Load test data, for instance
##
data("dTest6")
##
## Apply additive decomposition based on linear regression (default)
##
res<-decomposition_sma(dTest6, fp=4)
##
## Plot the decomposition components
##
plot_decomposition(res$Table)
##
## Apply multiplicative decomposition based on linear regression
##
res<-decomposition_sma(dTest6, fp=4, model="multiplicative")
##
## Plot the decomposition components
##
plot_decomposition(res$Table)

}
\seealso{
decomposition_lrtp_mult, decomposition_lrtp_add,
decomposition_sma_mult, decomposition_lrtp, plot_decomposition
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
