% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.R
\name{des}
\alias{des}
\title{Double Exponential Smoothing - Holt's algorithm}
\usage{
des(d, alpha = 0.3, beta = 0.2, future = 1, a0 = 0, b0 = NaN)
}
\arguments{
\item{{d}}{{numerical vector containing observed demand}}

\item{{alpha}}{{smoothing parameter for level update; the default value is 0.3}}

\item{{beta}}{{smoothing parameter for trend update; the default value 0.2}}

\item{{a0}}{{initialization level (previous forecast for first period); if it is not supplied, the first observed demand is selected}}

\item{{b0}}{{initialization trend; if it is not supplied, the difference between the first two observed demands is selected}}

\item{{future}}{{number of future periods to forecast}}
}
\value{
A data frame consisting of three columns:
\itemize{
\item\code{Demand} {it is the supplied \code{d}}
\item\code{Forecast} {it contains the predicted values for each period, icluding \code{future} periods}
\item\code{Level} {it contains the level updates produced during the process}
\item\code{Trend} {it contains the trend updates produced during the process}
\item\code{Error} {the residual values expressing the difference between \code{Demand} and \code{Forecast}}
}
}
\description{
Forecasting method of double exponential smoothing for time series model with additive trend and no seasonality.
}
\examples{
## Load test data, for example
data("dTest1")
## Run with default parameter values
des(d=dTest1)
## Run for 5 future periods
des(d=dTest1, future=5)
## Run with other parameter values
des(d=dTest1, future=5, alpha=0.6, beta=0.5)
## Run with non default initialization values for level and trend
des(d=dTest1, future=5, alpha=0.6, beta=0.5, a0=95, b0=85)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\seealso{
sma, wma, naive, ses, ma, tes, ddes, dtes, lrtp
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
