% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseasonize_add.R
\name{deseasonize_add}
\alias{deseasonize_add}
\title{Additively "de-seasonize" a demand time series (forecast)}
\usage{
deseasonize_add(d, fp = NaN)
}
\arguments{
\item{{d}}{{numerical vector of demand series (forecast)}}

\item{{fp}}{{seasonal frequency, for example 4 for quarters, 12 for months}}
}
\value{
A list contaíning
\itemize{
\item\code{Seasonal} {A vector that contains the seasonal components}
\item\code{Table} {A data frame consisting of the columns:
\itemize{
\item\code{Demand} {it is the supplied \code{d}}
\item\code{Deseasoned} {it contains the deseasoned values of \code{d}}
\item\code{Seasonality} {it containes the seasonal component to the corresponding term of \code{d}}
}}
\item\code{Fig1} {Contains a plot of time series versus seasonality}
\item\code{Fig2} {Contains a plot of the given time series vs the deseason values}
}
}
\description{
Computes the additive seasonal components of a time series (forecast) and de-seasonize it
by subtracting them from the given series. Clearly, it is assumed that the time
series model is additive.
}
\details{
Requires AMPrognosis::seasonal_add()
}
\examples{
## *** Example 1 ***
##
## Load test data, for example
##
data("dTest2")
##
## Run de-seasoning the series
##
res1 <- deseasonize_add(d=dTest2, fp=4)
##
## Plot the series vs its seasonality
##
print(res1$Fig1)
##
## Plot the series vs its deseasoned counter part
##
print(res1$Fig2)
##
## Run a forecast method for time series with trend on the de-seasoned series
##
res2 <- des(d=dTest2, future=5)
##
## Plot demand vs trend
##
plot_demand_vs_forecast(df=res2)
##
## Plot demand vs forecast
##
plot_demand_vs_forecast(df=data.frame(Demand=res2$Demand, Forecast=res2$Forecast))
##
## "Re-seasonize" forecast and plot it vs demand
##
sfactor <-res1$Seasonal
forecast <-res2$Forecast
df1 <- seasonize_add(forecast,sfactor,fp=4)
plot_demand_vs_forecast(df=data.frame(Demand=forecast, Forecast=df1$SeasonedForecast)
## 
## Compare with the next Example!!
##

## *** Example 2 (continues from the previous) ***
##
## Run a forecast method for time series with trend
##
res0 <- des(d=dTest2, future=5)
##
## Run de-seasoning the forecast
##
res <- deseasonize_add(d=res0$Forecast, fp=4)
##
## Plot the forecasted series vs its seasonality
##
print(res$Fig1)
##
## Plot the forecasted series vs its deseasoned counter part
##
print(res$Fig2)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\seealso{
deseasonize_mult
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
