% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtes_mult.R
\name{dtes_mult}
\alias{dtes_mult}
\title{Dumped Triple Exponential (Mixed) Multiplicative Smoothing - Gardner-McKenzie-Holt-Winters algorithm}
\usage{
dtes_mult(
  d,
  fp = NaN,
  alpha = 0.3,
  beta = 0.2,
  gamma = 0.1,
  phi = 1,
  future = 1,
  a0 = NaN,
  b0 = NaN
)
}
\arguments{
\item{{d}}{{numerical vector containing observed demand}}

\item{{fp}}{{seasonal frequency, for instance, 4 for quarters}}

\item{{alpha}}{{smoothing parameter for level update; the default value is 0.3}}

\item{{beta}}{{smoothing parameter for trend update; the default value 0.2}}

\item{{gamma}}{{smoothing parameter for seasonal update; the default value 0.1}}

\item{{phi}}{{dumping factor; default value of 1 results in classic Holt-Winters algorithm}}

\item{{a0}}{{initialization level (previous forecast for first period); if it is not supplied, the first observed demand is selected}}

\item{{b0}}{{initialization trend; if it is not supplied, the difference between the first two observed demands is selected}}

\item{{future}}{{number of future periods to forecast}}
}
\value{
A data frame consisting of three columns:
\itemize{
\item\code{Demand} {it is the supplied \code{d}}
\item\code{Forecast} {it contains the predicted values for each period, icluding \code{future} periods}
\item\code{Level} {it contains the level updates produced during the process}
\item\code{Trend} {it contains the trend updates produced during the process}
\item\code{Seasonal} {it contains the seasonal updates produced during the process}
\item\code{Error} {the residual values expressing the difference between \code{Demand} and \code{Forecast}}
}
A \code{NULL} is returned if \code{fp} is not supplied.
}
\description{
Forecasting method of triple exponential smoothing for time series model with
additive trend and multiplicative seasonality. A Gardner-McKenzie dumping factor
is used to dump future trend
}
\details{
Requires AMPrognosis::seasonal_mult()
}
\examples{
## Load test data, for example
data("dTest6")
## Run with default parameter values
dtes_mult(d=dTest6,fp=4)
## Run for 5 future periods with dumping factor 0.8
dtes_mult(d=dTest6, fp=4, future=5, phi=0.8)
## Run with other parameter values
dtes_mult(d=dTest6, fp=4, future=5, alpha=0.6, beta=0.5, gamma=0.2, phi=0.8)
## Run with non default initialization values for level and trend
dtes_mult(d=dTest6, fp=4, alpha=0.6, beta=0.5, gamma=0.2, phi=0.8, future=5, a0=400, b0=110)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\seealso{
sma, wma, naive, ses, ma, tes, ddes, dtes, lrtp, tes_add, tes_mult, dtes_add
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
