% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTest1.R
\docType{data}
\name{dTest1}
\alias{dTest1}
\title{Example of demand data}
\format{
Numerical vector
}
\usage{
dTest1
}
\description{
Data not based on any particular source
}
\seealso{
dTesT2 -- dTest10, dTest11fp4 -- dTest12fp4, dTest15 -- dTest16
}
\keyword{datasets}
