% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTest10.R
\docType{data}
\name{dTest10}
\alias{dTest10}
\title{Example of demand data}
\format{
Numerical vector
}
\usage{
dTest10
}
\description{
Data not based on any particular source
}
\seealso{
dTest1 -- dTest9, dTest11fp4 -- dTest12fp4, dTest15 -- dTest16
}
\keyword{datasets}
