% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTest13fp4.R
\docType{data}
\name{dTest13fp4}
\alias{dTest13fp4}
\title{Example of demand data with seasonality}
\format{
Numerical vector
}
\usage{
dTest13fp4
}
\description{
Data not based on any particular source.
}
\details{
Seasonality period is in quarters, that is, fp = 4
}
\seealso{
dTest1 -- dTest10, dTest11fp4 -- dTest12fp4, dTest14fp4, dTest15 -- dTesT16
}
\keyword{datasets}
