% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTest16.R
\docType{data}
\name{dTest16}
\alias{dTest16}
\title{Example of demand data}
\format{
Numerical vector
}
\usage{
dTest16
}
\description{
Data not based on any particular source
}
\seealso{
dTest1 -- dTest10, dTest11fp4 -- dTest12fp4, dTest15
}
\keyword{datasets}
