% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTest2.R
\docType{data}
\name{dTest2}
\alias{dTest2}
\title{Example of demand data}
\format{
Numerical vector
}
\usage{
dTest2
}
\description{
Data not based on any particular source
}
\seealso{
dTest1, dTesT3 -- dTest10, dTest11fp4 -- dTest12fp4, dTest15 -- dTest16
}
\keyword{datasets}
