% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTest3.R
\docType{data}
\name{dTest3}
\alias{dTest3}
\title{Example of demand data}
\format{
Numerical vector
}
\usage{
dTest3
}
\description{
Data not based on any particular source
}
\seealso{
dTest1, dTest2, dTesT4 -- dTest10, dTest11fp4 -- dTest12fp4, dTest15 -- dTest16
}
\keyword{datasets}
