% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTest4.R
\docType{data}
\name{dTest4}
\alias{dTest4}
\title{Example of demand data}
\format{
Numerical vector
}
\usage{
dTest4
}
\description{
Data not based on any particular source
}
\seealso{
dTest1 -- dTest3, dTest5 -- dTest10, dTest11fp4 -- dTest12fp4, dTest15 -- dTest16
}
\keyword{datasets}
