% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTest5.R
\docType{data}
\name{dTest5}
\alias{dTest5}
\title{Example of demand data}
\format{
Numerical vector
}
\usage{
dTest5
}
\description{
Data not based on any particular source
}
\seealso{
dTest1 -- dTest4, dTesT6 -- dTest10, dTest11fp4 -- dTest12fp4, dTest15 -- dTest16
}
\keyword{datasets}
