% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTest7.R
\docType{data}
\name{dTest7}
\alias{dTest7}
\title{Example of demand data}
\format{
Numerical vector
}
\usage{
dTest7
}
\description{
Data not based on any particular source
}
\seealso{
dTest1 -- dTest6, dTest8 -- dTest10, dTest11fp4 -- dTest12fp4, dTest15 -- dTest16
}
\keyword{datasets}
