% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTest8.R
\docType{data}
\name{dTest8}
\alias{dTest8}
\title{Example of demand data}
\format{
Numerical vector
}
\usage{
dTest8
}
\description{
Data not based on any particular source
}
\seealso{
dTest1 -- dTest7, dTest9 -- dTest10, dTest11fp4 -- dTest12fp4, dTest15 -- dTest16
}
\keyword{datasets}
