% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dTest9.R
\docType{data}
\name{dTest9}
\alias{dTest9}
\title{Example of demand data}
\format{
Numerical vector
}
\usage{
dTest9
}
\description{
Data not based on any particular source
}
\seealso{
dTest1 -- dTest8, dTest10, dTest11fp4 -- dTest12fp4, dTest15 -- dTest16
}
\keyword{datasets}
