% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Holt5256yr.R
\docType{data}
\name{Holt5256yr}
\alias{Holt5256yr}
\title{Example of annual orders received for the years 1952-1956}
\format{
Numerical vector
}
\usage{
Holt5256yr
}
\description{
Data is based on Table 7-2 in Holt et al. below
}
\references{
Holt, C. C., Modigliani, F., Muth, J. F., Simon, H. A., and Bonini, C. P. (1960) Plannin Production, Inventories, and Work Force, Prentice-Hall, Lodon.
}
\keyword{datasets}
