% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Holt5556mon.R
\docType{data}
\name{Holt5556mon}
\alias{Holt5556mon}
\title{Example of monthly orders received for the two year 1955-1956}
\format{
Numerical vector
}
\usage{
Holt5556mon
}
\description{
Data is based on Table 7-1 in Holt et al. below
}
\references{
Holt, C. C., Modigliani, F., Muth, J. F., Simon, H. A., and Bonini, C. P. (1960) Plannin Production, Inventories, and Work Force, Prentice-Hall, Lodon.
}
\keyword{datasets}
