% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterativeMAD.R
\name{iterativeMAD}
\alias{iterativeMAD}
\title{Iterative update of MAD and SD}
\usage{
iterativeMAD(df, alpha = 0.4, mad0 = 0)
}
\arguments{
\item{{df}}{{data frame containing at least columns \code{Demand} and \code{Forecast} and maybe \code{Error}. If \code{Error} is not present, it is created by the function.}}

\item{{alpha}}{{a smoothing parameter with value between 0 and 1}}

\item{{mad0}}{{initial MAD. If not given, the function will use the absolute value
of the first forecast error (residual)}}
}
\value{
A data frame wich extends the input \code{df} with two additional columns:
\itemize{
\item\code{MAD} {containing the successive updates of MAD}
\item\code{SD} {containing the successive updates of standard deviation}
}
}
\description{
Iterative update of MAD and estimation of standard deviation (SD) in accordance with
section 2.10.2 in Axsätter's book (see also section 6.3 in this book)
}
\examples{
## Perform a forecast
##
r0 <- des(dTest2, future=0)
##
##
## Generate iteratively MAD and SD approximations
##
iterativeMAD(r0)


}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
