% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtp.R
\name{lrtp}
\alias{lrtp}
\title{Linear regression - Trend projection}
\usage{
lrtp(d, future = 1)
}
\arguments{
\item{{d}}{{numerical vector containing observed demand}}

\item{{future}}{{number of future periods to forecast, default 1 period}}
}
\value{
A list containing three elements:
\itemize{
\item\code{Line} {a list containing the regression line coefficients; intercept, \code{a}, and slope, and \code{b}}
\item\code{Statistics} {a list containing the correlation measurements; correlation coefficent, \code{r}, and \code{r2}}
\item\code{Table} {a data frame consisting of three columns:
\itemize{
\item\code{Demand} {it is the supplied \code{d}}
\item\code{Forecast} {it contains the predicted values for each period, icluding \code{future} periods}
\item\code{Error} { its the residual values expressing the difference between \code{Demand} and \code{Forecast}}
}
}
}
}
\description{
Linear regression method to estimate the time series trend
and project it into the future to perform forecasting
}
\examples{
## Load test data, for example
data("dTest1")
## Run with default parameter values
lrtp(d=dTest1)
## Run for 5 future periods
lrtp(d=dTest1, future=5)


}
\seealso{
sma, wma, naive, ses, des, tes, ddes, dtes, ma
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
