% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtp_seasonize_mult.R
\name{lrtp_seasonize_mult}
\alias{lrtp_seasonize_mult}
\title{Multiplicatively "seasonize" a linear regression forecast}
\usage{
lrtp_seasonize_mult(d, fp = 4, future = 1)
}
\value{
A data frame consisting of the columns:
\itemize{
\item\code{Demand} {it is the supplied \code{d}}
\item\code{Forecast} {it contains the predicted values for each period, icluding \code{future} periods}
\item\code{Error} { its the residual values expressing the difference between \code{Demand} and \code{Forecast}}
\item\code{SeasonedForecast} {it contains the seasonally adjusted forecast}
\item\code{SeasonedError} {the residual values expressing the difference between \code{Demand} and \code{SeasonedForecast}}
}
}
\description{
Uses linear regression to project the trend into the future and determines the
multiplicative seasonal factors to seasonally adjust the regression forecast. The assumed time
series model is (mixed) multiplicative, i.e., additive trend and multiplicative seasonality.
}
\details{
Requires AMPrognosis::lrtp(),and  AMPrognosis::seasonal_mult()
}
\examples{
## Load test data, for example
data("dTest3")
## Run 
lrtp_seasonize_mult(d=dTest3, fp=4, future=6)


}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\seealso{
lrtp_season_add
}
