% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ma.R
\name{ma}
\alias{ma}
\title{Moving Average}
\usage{
ma(d, n = 3, future = 1)
}
\arguments{
\item{{d}}{{numerical vector containing observed demand}}

\item{{n}}{{order, i.e., number of periods to form the average, the default value is 3}}

\item{{future}}{{number of future periods to forecast}}
}
\value{
A data frame consisting of three columns:
\itemize{
\item\code{Demand} {it is the supplied \code{d}}
\item\code{Forecast} {it contains the predicted values for each period, icluding \code{future} periods}
\item\code{Error} { its the residual values expressing the difference between \code{Demand} and \code{Forecast}}
}
It returns NULL if the supplied order is incosistent with the length of d
}
\description{
Forecasting method of moving average
}
\examples{
## Load test data, for example
data("dTest1")
## Run with default parameter values
ma(d=dTest1)
## Run for 5 future periods
ma(d=dTest1, future=5)
## Run with moving average of order 5
ma(d=dTest1, n=5)
## Run with moving average of order 5 for 5 future periods
ma(d=dTest1, n=5, future=5)


}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\seealso{
sma, wma, naive, ses, des, tes, ddes, dtes, lrtp
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
