% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive.R
\name{naive}
\alias{naive}
\title{Naive forecasting procedures}
\usage{
naive(
  d,
  future = 1,
  fp = 4,
  method = c("naive", "mean", "trend", "drift", "seasonal")
)
}
\arguments{
\item{{d}}{{numerical vector containing observed demand}}

\item{{future}}{{number of future periods to forecast, default 1}}

\item{{fp}}{{needed if the \code{seasonal} procedure is pecified in order to provide the seasonal period frequency, dafault is 4, i.e., quarters}}

\item{{method}}{{specifies which naive procedure to be used. The alternatives are:
\itemize{
\item\code{naive} {simply repeat the last observed value}
\item\code{mean} {the mean value of all historic data}
\item\code{trend} {naive trend projection, i.e., linear interpolation between first and last observation, and extrapolation into the future (cf. linear regression)}
\item\code{drift} {the amount of change over time is set to be the average change seen in the historical data (cf. naive trend)}
\item\code{seasonal} {naive consideration of seasonality, i.e., repeat for each time period the last observed value for the corresponding period in the previous season}
}}}
}
\value{
A data frame consisting of three columns:
\itemize{
\item\code{Demand} {it is the supplied \code{d}}
\item\code{Forecast} {it contains the predicted values for each period, icluding \code{future} periods}
\item\code{Error} { its the residual values expressing the difference between \code{Demand} and \code{Forecast}}
}
It returns NULL if the supplied order or seasonality is incosistent with the length of d
}
\description{
Naive forecasting procedures (see the \code{method} parameter)
}
\details{
Method \code{drift} is just a slightly different implementation of \code{trend}
and therefore both will produce the same forecast. You can verify this by
listing the results of the two procedures side-by-side, for instance, like this:

\code{df1<-naive(d,future=3, method="trend")}

\code{df2<-naive(d,future=3, method="drift")}

\code{data.frame(df1$Forecast,df2$Forecast)}
}
\examples{
## Load test data, for example, 
data("dTest1")
##
## Run the default naive method for three future periods
##
naive(d=dTest1,future=3)
##
## Run the mean method for three future periods 
##
naive(d=dTest1,future=3, method="mean")
##
## Run the trend method for three future periods 
##
naive(d=dTest1,future=3, method="trend") 
##
## Run the drift method for three future periods
##
naive(d=dTest1,future=3, method="drift")
##
## Run the seasonal method for three future periods
##
naive(d=dTest1,future=3, method="seasonal", fp=4)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\seealso{
sma, wma, ma, ses, des, tes, ddes, dtes, lrtp
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
