% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_decomposition.R
\name{plot_decomposition}
\alias{plot_decomposition}
\title{Plots the decomposition of a demand time series}
\usage{
plot_decomposition(df, figtitle = NaN, figfil = NaN)
}
\arguments{
\item{{df}}{{adata frame containing the output of a decomposition function, that is, \code{Demand}, \code{DeTrended} demand, \code{Residual}, \code{Seasonality}, and \code{Trend}. \code{Seasonality} may be absent, of course,  in some cases.}}

\item{{figtitle}}{{Title of the plot. If not supplied, "Time Series Decomposition" is used}}

\item{{figfil}}{{Optionally, a file name to store the plot in A4 page. The extension determines the type of the plot; \code{.pdf} for PDF file, \code{.png} for PNG file, etc.}}
}
\value{
A figure in A4 page format containing plots of the decomposition components

#' @examples
\subsection{Load test data, for instance}{
}

\subsection{}{

data("dTest6")
}

\subsection{}{
}

\subsection{Apply additive decomposition based on linear regression}{
}

\subsection{}{

res<-decomposition_lrtp_add(dTest6, fp=4)
}

\subsection{}{
}

\subsection{Plot the decomposition components}{
}

\subsection{}{

plot_decomposition(res$Table)
}
}
\description{
Provided with the output of the decomposition functions, it
generates several plots in one figure to provide insight. It plots the figure on screen,
it can store it in a file, if a file name is specified, or it returns
it in variable, if specified. The variable can subsequently be
}
\details{
Requires the packages \code{ggplot2}, \code{gridExtra}, and \code{cowplot}
}
\seealso{
decomposition_lrtp_mult, decomposition_lrtp, decomposition_sma_add,
decomposition_sma_mult, decomposition_sma
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
