% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_demand_vs_forecast.R
\name{plot_demand_vs_forecast}
\alias{plot_demand_vs_forecast}
\title{Plot observed demand against a forecast (or against its trend)}
\usage{
plot_demand_vs_forecast(df)
}
\arguments{
\item{{df}}{{a data frame containing a column \code{Demand} and a column \code{Trend} or \code{Forecast}.
Optionally, it may contain a column \code{SeasonedForecast}}. Typically, the data frame is the resulting
output of other package funnctions such as \code{ma}, \code{sma}, \code{lrpt}, \code{ses}, etc.}
}
\value{
A plot of \code{Demand} vs \code{Forecast} (or \code{Trend}), and, possibly, vs \code{SeasonedForecast}.
The plot can be stored in a variable and displayed by printing the variable
}
\description{
The function plots the demand against estimated forecst or trend
}
\details{
Requires package \code{ggplot2}
}
\examples{
## Load test data, for example
data("dTest1")
## Run a forecast procedure, for instance
df0 <- ma(d=dTest1, n=5, future=5)
## Plot Demand vs Forecast
p1<-plot_demand_vs_forecast(df0)
print(p1)
## Estimate the time series demand
res <- lrtp(d=dTest1)
## Plot Demand vs Trend
p2 <- plot_demand_vs_forecast(df=res$Table)
print(p2)
## or plot without storing in a variable
plot_demand_vs_forecast(df=res$Table)


}
