% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_mult.R
\name{seasonal_mult}
\alias{seasonal_mult}
\title{Multiplicative Seasonality}
\usage{
seasonal_mult(d, fp)
}
\arguments{
\item{{d}}{{numerical vector containing observed demand}}

\item{{pf}}{{the seasonal periodicity (frequency), e.g., 4 for the quarters, 12 for the months of a year}}
}
\value{
A numerical vector containing the multiplicative seasonal factors
}
\description{
Computes the multiplicative seasonal factors (indices) of a time series
}
\examples{
## Load test data, for example
data("dTest3")
## Run 
seasonal_mult(d=dTest3, fp=4)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\seealso{
seasonal_add
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
