% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonize_mult.R
\name{seasonize_mult}
\alias{seasonize_mult}
\title{Multiplicatively "re-seasonize" a demand time series (forecast)}
\usage{
seasonize_mult(forecast, sfactor, fp = NAN)
}
\arguments{
\item{{forecast}}{{the demand series (forecast) to be re-seasonized}}

\item{{sfactor}}{{the additive seasonal components}}

\item{{fp}}{{the seasonal frequency, for example 4 for quarters, 12 for months}}
}
\value{
A data frame wit two columns:
\itemize{
\item\code{Forecast} {the given time series}
\item\code{SeasonedForecast} {the additively seasonally adjusted series}
}
}
\description{
Given the multiplicative seasonal factors of a de-seasoned time series (forecast), re-seasonize it
by multiplying them to the given series. Clearly, it is assumed that the time
series model is multiplicative.
}
\examples{
## *** Example 1 ***
##
## Load test data, for example
##
data("dTest2")
## Run de-seasoning the series
##
res1 <- deseasonize_mult(d=dTest2, fp=4)
##
## Run a forecast method for time series with trend on the de-seasoned series
##
res2 <- des(d=dTest2, future=5)
##
## "Re-seasonize" the forecast 
##
sfactor <-res1$Seasonal
forecast <-res2$Forecast
df1 <- seasonize_mult(forecast,sfactor,fp=4)
##
## and plot it vs demand
##
plot_demand_vs_forecast(df=data.frame(Demand=forecast, Forecast=df1$SeasonedForecast))
## 

}
\seealso{
seasonize_add
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
