% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.R
\name{ses}
\alias{ses}
\title{Simple Exponential Smoothing}
\usage{
ses(d, alpha = 0.3, future = 1, f0 = 0)
}
\arguments{
\item{{d}}{{numerical vector containing observed demand}}

\item{{alpha}}{{smoothing parameter between 0 and 1, default value is 0.3}}

\item{{f0}}{{initialization forecast (previous forecast for first period); if not supplied, the oldest observed demand is selected}}

\item{{future}}{{number of future periods to forecast}}
}
\value{
A data frame consisting of three columns:
\itemize{
\item\code{Demand} {it is the supplied \code{d}}
\item\code{Forecast} {it contains the predicted values for each period, icluding \code{future} periods}
\item\code{Error} {the residual values expressing the difference between \code{Demand} and \code{Forecast}}
}
It returns NULL if the supplied alpha is not between 0 and 1
}
\description{
Forecasting method of simple exponential smoothing
}
\examples{
## Load test data, for example
data("dTest1")
## Run with default parameter values
ses(d=dTest1)
## Run for 5 future periods
ses(d=dTest1, future=5)
## Run with smoothing parameter 0.5
ses(d=dTest1, alpha=0.5)
## Run with smoothing parameter 0.5 for 5 future periods
ses(d=dTest1, alpha=0.5, future=5)
## Run with initialization forecast
ses(d=dTest1, alpha=0.5, future=5, f0=76)


}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\seealso{
sma, wma, naive, ma, des, tes, ddes, dtes, lrtp
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
