% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma.R
\name{sma}
\alias{sma}
\title{Symmetric Moving Average}
\usage{
sma(d, n = 3)
}
\arguments{
\item{{d}}{{numerical vector containing observed demand}}

\item{{n}}{{order, i.e., number of periods to form the average, the default value is 3}}
}
\value{
A data frame consisting of three columns:
\itemize{
\item\code{Demand} {it is the supplied \code{d}}
\item\code{Trend} {it contains the predicted values for each period}
}
It returns NULL if the supplied order is incosistent with the length of d
}
\description{
The method of symmetric moving average to estimate the time series trend.
}
\details{
Note that the function does not produce  future forecast and that
it does not determine the error (residual) \code{Demand - Trend} either,
but this is easily
produced from the returned result, for instance,  as follows:

\code{df <- sma(d=dTest1)}

\code{error <- df$Demand - df$Trend}
}
\examples{
## Load test data, for example
data("dTest1")
## Run with default parameter values
sma(d=dTest1)
## Run with symmetric moving average of order 5
sma(d=dTest1, n=5)


}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\seealso{
ma, wma, naive, ses, des, tes, ddes, dtes, lrtp
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
