% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes.R
\name{tes}
\alias{tes}
\title{Triple Exponential Smoothing - Holt-Winters algorithm}
\usage{
tes(
  d,
  fp = NaN,
  alpha = 0.3,
  beta = 0.2,
  gamma = 0.1,
  future = 1,
  model = c("additive", "multiplicative"),
  a0 = 0,
  b0 = NaN
)
}
\arguments{
\item{{d}}{{numerical vector containing observed demand}}

\item{{fp}}{{seasonal frequency, for instance, 4 for quarters}}

\item{{alpha}}{{smoothing parameter for level update; the default value is 0.3}}

\item{{beta}}{{smoothing parameter for trend update; the default value 0.2}}

\item{{gamma}}{{smoothing parameter for seasonal update; the default value 0.1}}

\item{{a0}}{{initialization level (previous forecast for first period); if it is not supplied, the first observed demand is selected}}

\item{{b0}}{{initialization trend; if it is not supplied, the difference between the first two observed demands is selected}}

\item{{future}}{{number of future periods to forecast}}

\item{{model}}{{can be \code{"additive"} or \code{"multiplicative"}; the additive case is the default one}}
}
\value{
A data frame consisting of three columns:
\itemize{
\item\code{Demand} {it is the supplied \code{d}}
\item\code{Forecast} {it contains the predicted values for each period, icluding \code{future} periods}
\item\code{Level} {it contains the level updates produced during the process}
\item\code{Trend} {it contains the trend updates produced during the process}
\item\code{Seasonal} {it contains the seasonal updates produced during the process}
\item\code{Error} {the residual values expressing the difference between \code{Demand} and \code{Forecast}}
}
A \code{NULL} is returned if \code{fp} is not supplied or if \code{method} is assigned wrong value.
}
\description{
Forecasting method of triple exponential smoothing for time series.
Provides a common interface to  both additive and (mixed) multiplicative model
}
\details{
Requires AMPrognosis::tes_add(), AMPrognosis::tes_add(),
AMPrognosis::seasonal_mult(), AMPrognosis::seasonal_add()
}
\examples{
## Load test data, for example
data("dTest6")
## Run for additive model (default)
dtes(d=dTest6,fp=4)
## Run for multiplicative model
tes(d=dTest6,fp=4, model="multiplicative")

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\seealso{
sma, wma, naive, ses, ma, ddes, dtes, lrtp, tes_add, tes_mult,
dtes_add, dtes_mult
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
