% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes_add.R
\name{tes_add}
\alias{tes_add}
\title{Triple Exponential Additive Smoothing - Holt-Winters algorithm}
\usage{
tes_add(
  d,
  fp = NaN,
  alpha = 0.3,
  beta = 0.2,
  gamma = 0.1,
  future = 1,
  a0 = NaN,
  b0 = NaN
)
}
\arguments{
\item{{d}}{{numerical vector containing observed demand}}

\item{{fp}}{{seasonal frequency, for instance, 4 for quarters}}

\item{{alpha}}{{smoothing parameter for level update; the default value is 0.3}}

\item{{beta}}{{smoothing parameter for trend update; the default value 0.2}}

\item{{gamma}}{{smoothing parameter for seasonal update; the default value 0.1}}

\item{{a0}}{{initialization level (previous forecast for first period); if it is not supplied, the first observed demand is selected}}

\item{{b0}}{{initialization trend; if it is not supplied, the difference between the first two observed demands is selected}}

\item{{future}}{{number of future periods to forecast}}
}
\value{
A data frame consisting of three columns:
\itemize{
\item\code{Demand} {it is the supplied \code{d}}
\item\code{Forecast} {it contains the predicted values for each period, icluding \code{future} periods}
\item\code{Level} {it contains the level updates produced during the process}
\item\code{Trend} {it contains the trend updates produced during the process}
\item\code{Seasonal} {it contains the seasonal updates produced during the process}
\item\code{Error} {the residual values expressing the difference between \code{Demand} and \code{Forecast}}
}
A \code{NULL} is returned if \code{fp} is not supplied.
}
\description{
Forecasting method of triple exponential smoothing for time series model with
additive trend and seasonality.
}
\details{
Requires AMPrognosis::seasonal_add()
}
\examples{
## Load test data, for example
data("dTest6")
## Run with default parameter values
tes_add(d=dTest6,fp=4)
## Run for 5 future periods
tes_add(d=dTest6, fp=4, future=5)
## Run with other parameter values
tes_add(d=dTest6, fp=4, future=5, alpha=0.6, beta=0.5, gamma=0.2)
## Run with non default initialization values for level and trend
tes_add(d=dTest6, fp=4, alpha=0.6, beta=0.5, gamma=0.2, future=5, a0=400, b0=110)

}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\seealso{
sma, wma, naive, ses, ma, tes, ddes, dtes, lrtp, tes_mult, dtes_add, dtes_mult
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
