% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wma.R
\name{wma}
\alias{wma}
\title{Weighted Moving Average}
\usage{
wma(d, w = c(1/3, 1/3, 1/3), future = 1)
}
\arguments{
\item{{d}}{{numerical vector containing observed demand}}

\item{{w}}{{numerical vector of weights, which determines the order, i.e., number of periods to form the weighted average, the default value is c(1/3,1/3,1/3)}}

\item{{future}}{{number of future periods to forecast}}
}
\value{
A data frame consisting of three columns:
\itemize{
\item\code{Demand} {it is the supplied \code{d}}
\item\code{Forecast} {it contains the predicted values for each period, icluding \code{future} periods}
\item\code{Error} { its the residual values expressing the difference between \code{Demand} and \code{Forecast}}
}
It returns NULL if the supplied weights don't add to 1 or imply inappropriate order
}
\description{
Forecasting method of weighted moving average
}
\examples{
## Load test data, for example
data("dTest1")
## Run with default parameter values
wma(d=dTest1)
## Run for 5 future periods
wma(d=dTest1, future=5)
## Run with moving average of order 5
wma(d=dTest1, w=c(1/5,1/5,1/5,1/5,1/5) )
## Run with moving average of order 5 for 5 future periods
wma(d=dTest1, w=c(1/5,1/5,1/5,1/5,1/5), future=5)


}
\references{
\itemize{
\item Axsäter, S. (2015) Inventory Control, 3d Edition, Springer
\item Snyder, L. V. and Shen, Z.-J., M. (2011) Fundamentals of Supply Chain Theory, Wiley
\item Cowpertwait, P. S. and Metcalfe, A. V. (2009) Introductory Time Series with R, Use R!, Springer
\item Hyndman, R. J. and Athanasopoulos, G. (2021) Forecasting Principles and Practive, 3d Edition, OTexts
}
}
\seealso{
sma, ma, naive, ses, des, tes, ddes, dtes, lrtp
}
\author{
Athanasios Migdalas, \email{athmig59@gmail.com}
}
